moved interfact props to the correct folder
