+++ src/app/software-projects/interfaceProps.ts
